# openapi.yaml
openapi: 3.0.3
info:
  title: XORFS HTTP API
  version: 1.0.0
  description: |
    HTTP REST API for the XORFS distributed filesystem. Provides endpoints to upload, download, list and delete files.
servers:
  - url: http://localhost:8080
paths:
  /files:
    get:
      summary: List files for a client
      parameters:
        - name: clientId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of filenames owned by the client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListFilesResponse'
        '4XX':
          description: Client error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Upload a file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - clientId
              properties:
                clientId:
                  type: integer
                  format: int64
                filename:
                  type: string
                  description: Optional filename override
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: File uploaded and allocated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '4XX':
          description: Upload failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /files/{clientId}/{filename}:
    parameters:
      - name: clientId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: filename
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Download a file
      responses:
        '200':
          description: File binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '4XX':
          description: Download failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a file
      responses:
        '200':
          description: File deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '4XX':
          description: Deletion failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    StandardResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            error_code:
              type: string
    UploadResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            clientId:
              type: integer
              format: int64
            filename:
              type: string
            stripes:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/StripeMetadata'
    DeleteResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            deleted_chunks:
              type: integer
    ListFilesResponse:
      type: object
      properties:
        filenames:
          type: array
          items:
            type: string
    StripeMetadata:
      type: object
      properties:
        stripe_num:
          type: integer
        chunk_ids:
          type: array
          items:
            type: string
        servers:
          type: array
          items:
            type: string
security: []